+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Clearing end device events list in the Console.
+
 ### Security
 
 ## [3.8.1] - 2020-05-11